Change the file structure of the pages
