+++ src/App.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import './App.css';
 import { Route, Routes } from 'react-router-dom';
 
-import { LoginPage } from 'pages/login-page';
-import { FrontPage } from 'pages/front-page';
+import { LoginPage } from './pages/login-page';
+import { FrontPage } from './pages/front-page';
 
 function App() {
   return (
