+++ src/pages/front-page/style.css
@@ -1,0 +1,18 @@
+.FrontPage {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  min-width: 500px;
+}
+
+.FrontPage__table {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 80%;
+}
+
+.FrontPage__forms {
+  width: 200px;
+}
